requirements.txt
# Django 核心
Django==4.2.7
django-jet-reboot==1.3.3

# 数据库
django-extensions==3.2.3

# Ansible
ansible==8.5.0
ansible-core==2.15.5

# 异步任务
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# 工具库
paramiko==3.3.1
jinja2==3.1.2
pyyaml==6.0.1
python-dotenv==1.0.0

# CSV 处理
pandas==2.1.3

# API
djangorestframework==3.14.0

# 差异对比
difflib2==0.1

# 时区
pytz==2023.3

# 开发工具
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

